backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
email-validator
alembic
pytest
httpx
flake8
black
ruff
pytest-cov
bcrypt
pydantic-settings
faker
isort
pylint
pre-commit
pydantic_settings
PyJWT
slowapi
passlib
psutil
